requirements.txt
pyyaml

##-e .
